LPD-78127 We are encouraging consumers to use just the slice they need when interacting with the atom, search or filter right now but might be others in the future, so it does not make sense to expose the whole getFDSAtom publicly
