+++ projects/js-toolkit/packages/js-api/data-set/index.ts
@@ -6,10 +6,11 @@
 /**
  * Public extension API for the Frontend Data Set (FDS) widget.
  *
- * `getFDSAtom` resolves a typed `Atom<FDSState>` once the data set has
- * registered its state under `Liferay.State`, polling until the atom
- * appears or the timeout elapses — useful when client code needs to
- * read or react to filter and search state from outside the widget.
+ * Consumers interact with FDS state through opinionated subscription
+ * helpers — one per blessed slice (`subscribeSearch`,
+ * `subscribeFilters`). New slices are added by exposing a new helper
+ * from this module; consumers do not reach for the underlying atom or
+ * selector directly.
  *
  * The remaining interfaces are the contracts that custom cell renderers
  * and filters implement to integrate with FDS: HTML element builders
@@ -19,80 +20,12 @@
 
 // Frontend data set state
 
-export type Atom<T> = Liferay.State.Atom<T>;
-export type Selector<T> = Liferay.State.Selector<T>;
-
-interface FDSFilterState {
-	active?: boolean;
-	id: string;
-	odataFilterString?: string;
-	selectedData?: Record<string, unknown>;
-}
-
 export interface FDSState {
-	filters: Array<FDSFilterState>;
 	search: {query: string};
 }
 
-export {dataSetSearch} from './dataset-search';
-export type {DataSetSearch} from './dataset-search';
-
-const DEFAULT_TIMEOUT = 5000;
-const DEFAULT_INTERVAL = 100;
-
-export function getFDSAtom(
-	id: string,
-	options?: {interval?: number; timeout?: number}
-): Promise<Atom<FDSState>> {
-	const timeout = options?.timeout ?? DEFAULT_TIMEOUT;
-	const interval = options?.interval ?? DEFAULT_INTERVAL;
-
-	const key = `${id}_fdsState`;
-
-	return new Promise((resolve, reject) => {
-		const existing =
-			Liferay.State.__unsafe__.getAtomOrSelectorKey(key);
-
-		if (existing) {
-			return resolve(existing as Atom<FDSState>);
-		}
-
-		const startTime = Date.now();
-
-		const poll = setInterval(() => {
-			const atom =
-				Liferay.State.__unsafe__.getAtomOrSelectorKey(key);
-
-			if (atom) {
-				clearInterval(poll);
-
-				return resolve(atom as Atom<FDSState>);
-			}
-
-			if (Date.now() - startTime >= timeout) {
-				clearInterval(poll);
-
-				reject(
-					new Error(
-						`FDS atom "${key}" was not found within ${timeout}ms`
-					)
-				);
-			}
-		}, interval);
-	});
-}
-
-export function getOrCreateSelector<T>(
-	key: string,
-	deriveValue: (get: Liferay.State.Getter) => T
-): Selector<T> {
-	const existing = Liferay.State.__unsafe__.getAtomOrSelectorKey(key);
-
-	return (
-		(existing as Selector<T> | null) ??
-		Liferay.State.selector<T>(key, deriveValue)
-	);
-}
+export {subscribeSearch} from './dataset-search';
+export type {SearchSubscription} from './dataset-search';
 
 // Frontend data set cell renderer
